Make Functions library file for cleanup
